Various enhancements to remark-images-download
I tried to make it more secure and stable: I believe dowloading files
from unknown servers should be done with as much care as processing
untrusted user input. I added a timeout (`httpRequestTimeout` which
is 2 seconds by default) and ensured that too long streams are
canceled: Without `Content-Length` header and with
`Transfer-Encoding: chunked`, one can easily fill the storage of a
zmarkdown server.

remark-images-download downloads images in parallel but wants to
check the sum of `Content-Length` headers is smaller than
`dirSizeLimit`: I think the best way to do this is to fetch
headers first, then sum and compare `Content-Length`s and download
the bodies. We can do all of this in parallel but we need some kind
of “barrier” for the “sum and compare” step in between. I tried
to make this parallelism stuff more explicit.

Identical URLs are downloaded only once. I added a few dummy query
parameters in the tests to make URLs unique when needed.

Download folders are always removed when empty.

I didn’t manage to make all the streaming stuff work with the `request`
package so I used lower-level Node.js APIs instead.

And of course, I added tests.