⚠️ Update remark-numbered-foonotes
**This commit only convert new identifiers to string.**
Identifiers used to be strings but in the previous version of this plugin,
identifiers were converted to numbers.

Now the type is preserved.

Adapt the plugin to the commit of mdast-util-to-hast#fd38c454
See the issue: https://github.com/syntax-tree/mdast-util-to-hast/pull/32

It updates the order of footnotes to a more coherent one.
But identifiers no longer allowed to be number.