[remark-images-download] [zmarkdown] Fix minor LFI
A minor Local File Inclusion vulnerability has been found in
remark-download-images, which allowed for images with a known path on
the host machine to be included inside a LaTeX document.

To prevent it, a new option has been created that allow to replace
invalid pathes with a default image instead of linking the image on the
host directly.