[rebber-plugins] [zmarkdown] Fix a critical RCE
A Remote Command Execution vulnerability was found in the rebber-plugins
module, which allowed
execution of arbitrary commands. The reported problem came
from math, which could contain malicious LaTeX.

The mitigation involves adding a module
that allows filtering the authorized
math commands to a given list.
By default, the list used is that of KaTeX.