Add remark-abbr to tests and re-record snapshots
The only difference in the snapshots seems to be the presence of `\\`
backslashes before quote characters. I'm a little confused by the
output, not being all that familiar with jest, but I suspect this is
some minor difference in the way that rehype-stringify works and that we
don't need to worry about it?