Add micromark-extension-abbr
This is basically the same as the version in
https://github.com/richardTowers/remark-abbr, except I've converted the
tests from node:test to jest to be consistent with the rest of
zmarkdown.

This only exports a syntax extension - there's no HTML extension
(because micromark doesn't currently support the hooks we'd need to
implement one - see https://github.com/orgs/micromark/discussions/181
for more detail). It's only intended to be used in remark-abbr.